remove HashResource in favor of generated hash methods
